apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
asyncpg==0.29.0
sqlalchemy==2.0.36
redis==5.0.8
celery==5.4.0
pydantic==2.9.1
pydantic-settings==2.5.2
python-slugify==8.0.4
python-dotenv==1.0.1
alembic==1.13.2
psycopg2-binary==2.9.9
pytest==8.3.2
httpx==0.27.2
pytest-asyncio==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
